+++ test/syntax.jl
@@ -287,7 +287,7 @@ call2(f,x,y) = f(x,y)
 @test (call2(42,1) do x,y; x+y+1 end) == 44
 
 # definitions using comparison syntax
-let a⊂b = reduce(&, x ∈ b for x in a) && length(b)>length(a)
+let a⊂b = all(x ∈ b for x in a) && length(b)>length(a)
     @test [1,2] ⊂ [1,2,3,4]
     @test !([1,2] ⊂ [1,3,4])
     @test !([1,2] ⊂ [1,2])